Fix the version settings due to new poky
Due to update of poky, old method to setting version for
linux-libc-headers and various gcc packages do not work anymore.
This causes wrong version in these packages.
This system is still buildable, but the cross-compile SDK will not work.

This patch fixes these old version settings.
Also remove compatible_machine in linux-libc-headers because it will
prevent the nativesdk package

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>