recipes-kernel:add uvcs-kernel-module for LTS4.4
CHANGE by Huan Thai <huan.thai.xw@rvc.renesas.com>:
These are issues when porting from 3.10 to 4.4
    Issue 1: Can not request interrupt for HW IP (VCP)
    Issue 2: Can not map HW IP register (VPC)
These is a difference in register platform device for the driver:
    - The old driver use kernel API (platform_device_register())
to register it device to kernel,
    - The new driver, device binding to driver is register by the new way
that use the device info from <target-board>.dtb file.

vpc1, vpc0xy, and vpc1xy are added as reserve only vcp0, vpc0 and prr is used.
The prr is used for getting LSI information which tell the driver
what kind of board it running

Signed-off-by: Long Dang <long.dang.aj@rvc.renesas.com>