Fix error when apply patch to MMP Full package
Below patch can apply to MMP trial package (has time limit).
But it cannot apply to MMP Full package (due to few differences with trial).

Fix the conflict parts so that the patch can be applied to both version.
(The conflict parts are actually just additional modifications, not relate
to main feature of this patch, so we can remove them safely)

Signed-off-by: HungTran <hung.tran.jy@rvc.renesas.com>