qt5: qtdeclarative: workarround for crash issue on iwg21m
I was debuging the issue crash when use text with animation effect on
iwg21m and found that the issue just happend when qt passed an odd-width
to resizeTexture().
This commit temporarily plus 1 to the width if it is odd to prevent the issue.
Further testing and investigating are needed

Signed-off-by: Binh Nguyen <binh.nguyen.uw@rvc.renesas.com>