SDK,cip-core/glibc: Fix nativesdk packages
Currently glibc packages in recipes-cip-core
(and also in its original source, meta-debian)
do not provide nativesdk correctly for packages
such as glibc-initial and glibc-mtrace.

Thus, when user build nativesdk, a different glibc
will be used for building nativesdk-glibc-initial,
nativesdk-glibc-mtrace, and some others.
Fix it in this commit.

Also fix a compile error due to wrong GCC for nativesdk

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>