gstreamer: Improve waylandsink performance in non-dmabuf mode
If input buffer is not dmabuf, SHM mode will be used.
However, in this mode buffers are not contiguous, thus
hardware process element like vspmfilter need to use
their own buffers, and then copy to SHM buffers later.
This causes the low performance.

This commit add KMS mode and use it as default in
this case. KMS buffers are contiguous so hardware
process can use them directly, thus improve performance.

Note that there is no change in DMABUF mode.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>