gstreamer: fix error when playing a problem video
A special video that have I frame at frame 30 cannot
be decoded. This issue only happens when
stream-format is byte-stream.
In older environment that video can be played with
stream-format=avc, but avc is removed since it
is just a work-around, not supported in OMX library.

This commit improves gst-omx and allow it to play
that file.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>