sdk: fix missing libraries and headers in sdk
Before this change, executable files in SDK cannot be run
normally after installed to Ubuntu 16.04 because system
cannot find correct libraries (not happen if install to Ubuntu 14.04)
The issue happens because rpath info of each executable files
is not correct. Possible causes is the version of glibc we are using.

This commit adds a post-processing script that automaticaly run
after intall SDK. This script will adjust the rpath information
and allow the binaries in SDK run normally.

Besides, this commit also add some header files which are necessary
for developement.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>