Weston: Fix issue that Weston cannot boot with media devices plugged in
The default device nodes created by VSP driver are /dev/media0 and
/dev/media1. If a media device (for example, a USB Camera) is plugged in
prior to system boot, the VSP device nodes might be changed to
/dev/media1, /dev/media2, and so on. This could make Weston confused and
unable to boot up.

This patch is to create symbolic links that can dynamically points to
device nodes of VSP driver in all situations. For that, Weston always
read/write on the right device nodes.

Authored-by: Duy Nguyen <duynt26@fsoft.com.vn>
Authored-by: Danh Duong <danhdc1@fsoft.com.vn>
Signed-off-by: Hao Bui <hao.bui.yg@renesas.com>