gstreamer-plugins-bad: Close drm file description and dma buffer after using
When playing many times of the pipeline with the "waylandsink" element, memory leakage
occured because the drm file description and dma buffer are not free correctly and
then memory is not enough to play the pipeline with wayland kms.

This commit is to close drm file description and dma buffer correcty after using to
prevent of causing memory leak in DU driver.

Signed-off-by: Le Hoang Minh Thuan <thuanlhm@fsoft.com.vn>