ltp:async_handler(_jk): adjust pass criteria and apply warm_up phase
When running async_handler test on rt system, we got high latency of the
pthread_cond_signal on RZ/G1E and RZ/G1C boards with dual cortex a7 CPU.
We also experienced abnormal value of the first latency results in loops
due to lacking of warm-up periods.

This patch applies a warm-up phase of 100us and adjust the pass criteria
to make it more proper for these boards.

For the async_handler_jk test, we change the starting measurement point
(marked by rt_gettime()) to exclude the time of configuration of server
which should not be counted into latency values.

Signed-off-by: Le Hoang Minh Thuan <thuanlhm@fsoft.com.vn>
Signed-off-by: Danh Duong <danhdc1@fsoft.com.vn>
Signed-off-by: Hao Bui <hao.bui.yg@renesas.com>