ARM: dts: iwg20d-g1n: restructure device tree following kernel changes
RZ/G1M and RZ/G1N SoCs can work with the same carrier board and camera
daughter board. This patch rename RZ/G1N DT to make it consistent to
that of RZ/G1M and better reuse the common parts of the two.

The final DT architecture to describe the board is now:
    r8a7744-iwg20d-q7.dts           # Carrier Board + SoM
    ├── r8a7744-iwg20m.dtsi         # SoM
    │   └── r8a7744.dtsi            # SoC
    └── iwg20d-q7-common.dtsi       # Carrier Board

Signed-off-by: Hao Bui <hao.bui.yg@renesas.com>