gst-bad: waylandsink: Fix memory leak in kms mode
The buffers in kmspool are not freed due to trouble in refcount.

Due to update of Linux kernel, a bug in drmPrimeHandleToFD was fixed and
now this functions always increases refcount of coresponding CMA object.
In user side, the returned fd must be closed after use to counter this
refcount-increasing. It is done in this patch.

Besides, also fix an harmless warning message due to misisng init
function:
"GStreamer-CRITICAL **: Registering meta implementation 'GstWlKmsMeta'
without init function".

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>
Signed-off-by: Hao Bui <hao.bui.yg@renesas.com>