templates: gecko: fix build for gecko due to openssl and openssh version
Currently Gecko is failed to build due to recently added CIP Core Buster
and upgraded ca-certificates packages.
If Buster-limited option is enabled for CIP Core Buster, openssl version
will be forced to 1.1.x.
This causes build error with Gecko.

Fix it by removing forcevariable version for openssl.
With this change, system will use old openssl version same as before.
Also change openssh version for matching with openssl.
(Still use other packages from CIP Core Buster, besides openssl and
openssh).

Besides, note that it is necessary to add a patch to ca-certificates.
This patch should be applied automatically for old openssl version, but
it is not because Buster option is enabled.

Signed-off-by: Quynh Nguyen <quynh.nguyen.xb@renesas.com>