recipes-debian: buster: binutils: Adjust TEXT_START_ADDR for RZ/Five
With applications compiled with a static flag the virtual address for
the applications may fall within H’0_0003_0000 - H’0_0004_FFFF where the
ILM and DLM blocks of the AX45MP exist.

The MMU won't trigger page faults when the virtual address falls in the
range of AX45MP local memory. So to make sure statically compiled
applications run successfully, adjust the TEXT_START_ADDR to 0x50000.

Signed-off-by: Lad Prabhakar <prabhakar.mahadev-lad.rj@bp.renesas.com>
Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>