docs: template: conf: Enable wayland and disable x11 in nativesdk
This patch enables wayland and disables x11 in nativesdk.
It applies for legacy build procedure which does not support RZ/G1 Series,
RZ/V2M, and RZ/V2MA.

Without it, nativesdk-packagegroup-sdk-host package will not add
nativesdk-wayland and nativesdk-wayland-dev.

Therefore, SDK of either core-image-weston or core-image-qt will not contain
wayland libaries, headers, or wayland-scanner tool for host machine and will
not be able to build applications which rely on wayland-scanner as below:

  | user@ubuntu:~/renesas/demo-camera-image-overlay$ source /opt/poky/3.1.21/\
  | environment-setup-aarch64-poky-linux
  |
  | user@ubuntu:~/renesas/demo-camera-image-overlay$ make wayland-scanner code \
  | /opt/poky/3.1.21/sysroots/aarch64-poky-linux/usr/share/wayland-protocols/\
  | stable/xdg-shell/xdg-shell.xml common/src/xdg-shell-protocol.c
  |
  | make: wayland-scanner: Command not found
  | make: *** [Makefile:66: common/src/xdg-shell-protocol.c] Error 127

Signed-off-by: Loc Pham <loc.pham.xh@renesas.com>