wic: Update wks files to improve usage for WIC image
Current setting of WIC image is:
   - Partition 1: srec files for SCIF download
   - Partition 2: rootfs + Linux Image and devicetree

This structure does not contain bootloader, and cannot be used
for eMMC or eSD boot easily.

This change adds:
   - Add an unformat area before partition 1 to contain bootloader
   - Remove mount points of partition 1 and 2
   - Update WKS file names and descriptions
   - Change default WKS file

The bootloader added can be used in eSD boot.
(for eMMC, user have to manually write bootloader to boot0 partition).
Also, by removing the mount points, no entry will be added to fstab,
thus same rootfs can be used for mmc0 or mmc1

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>