wic: rewrite wic dependency to fix build error in rzv2m and rzfive
Previously do_wic is set to depends on u-boot and trusted-firmware-a.
This is cofrrect for rzg2 and rzg2l, rzv2l.
But rzv2m and rzfive do not have trusted-firmware-a, thus have build error.

Fix this by rewriting the dependency.
Now move the dependent list to machine config file, next to IMAGE_BOOT_FILES.
With this, the dependent list for do_wic can be written correctly based on
the files used in IMAGE_BOOT_FILES.

Also, previously IMAGE_BOOT_FILES for rzv2m missed some loader files.
These files are provided by another layer, so have to add them with condition.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>