rz-bsp: weston: Revert "gl-renderer: Don't use TEXTURE_EXTERNAL for multi-planar formats"
Reason: GL_TEXTURE_2D on Mali graphics did not support YUV formats. As a
result, a black pop-up window will be displayed if you try to use
GL_TEXTURE_2D with YUV formats. It is necessary to revert this commit to use
GL_TEXTURE_EXTERNAL_OES with YUV formats, that is supported.

Signed-off-by: trungvanle <trung.le.xk@renesas.com>