machine.conf : Fix dependencies of do_image_wic
The variable "WIC_INPUT_DEPENDS" is not provided on scarthgap recipe.
This is Renesas unique variable, so it should not be used.
On dunfell recipe, this variable is reassigned to "do_image_wic[depends]"
to fourcely remake wic image if it's component were changed.
"WKS_FILE_DEPENDS" can be used instead, but it makes to run
do_rootfs again.

So I changed to use "do_image_wic[depends]" directly to reflect
these dependencies.

Signed-off-by: Yoshihisa Yamamoto <yoshihisa.yamamoto.xn@renesas.com>
Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>