kas: Add support to build with proprietary libraries
Add new Kas configuration options so that a build can include
meta-rz-graphics and/or meta-rz-codecs from the meta-rz-features
package.

These options can be selected in the Kas menu or by appending them to the
Kas build command. For example:
  ./kas-container build \
    "kas/base.yml
    :kas/images/core-image-weston.yml
    :kas/machines/smarc-rzg3e.yml
    :kas/features/graphics.yml"

Note that both of these layers must be downloaded from the Renesas website
separately and extracted into the working directory. For example:

meta-renesas
├── COPYING.MIT
├── kas
├── kas-container
├── Kconfig
├── meta-arm
│   ├── meta-arm
│   └── meta-arm-toolchain
├── meta-openembedded
│   ├── meta-multimedia
│   ├── meta-oe
│   └── meta-python
├── meta-rz-boards
├── meta-rz-bsp
├── meta-rz-features
│   ├── meta-rz-codecs
│   └── meta-rz-graphics
└── poky
    ├── meta
    ├── meta-poky
    └── meta-yocto-bsp

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>