meta-rz-common: Modify DEPENDS/RDEPENDS of some packages
When building the core image, some packages such as kernel-module,
user-module of mmngr, vspm and gstreamer plugins are installed
via packagegroups.
Howerver, if we missed to install these packagegroups,
other packages which depends on them will be failed in runtime.
We should define RDEPENDS variable for some recipes so that they are
installed with their runtime dependencies, including:
- gst-plugins-bad, gst-plugins-good
- mmngr-user-module, mmngrbuf-user-module, vspmif-tp-user-
module and vspmif-user-module
- kernel-module-vspmif

Signed-off-by: Kim Quynh <quynh.nguyen.xr@renesas.com>