Modify DEPENDS/RDEPENDS of some packages
When building the core image, some packages such as kernel-
module, user-module of mmngr, vspm and gstreamer plugins
are installed via packagegroups. Howerver, if we missed to
install these packagegroups. Other packages which depends
on them will be failed in runtime. So, we should define
RDEPENDS variable for some recipes so that they are
installed with their runtime dependencies. Including:
- mmngr-user-module, mmngrbuf-user-module, vspmif-tp-
user-module and vspmif-user-module.
- kernel-module-vspmif
- gst-plugins-good

Signed-off-by: Kim Quynh <quynh.nguyen.xr@renesas.com>