rz-bsp: trusted-firmware-a: Remove redundant do_compile:prepend()
The custom do_compile:prepend() in trusted-firmware-a_2.10.bb was adding
BUILD_LDFLAGS to the fiptool Makefile. However, this step is already handled
by the do_compile() defined in trusted-firmware-a.inc from meta-arm,
which this recipe inherits.

Removing it avoids duplication and ensures maintenance is aligned with
the shared upstream logic.

Signed-off-by: Au Dang <au.dang.jx@renesas.com>